not-so-smart-contracts/cosmos/incorrect_getsigners/x/incorrectgetsigners/types/message_delegate_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"github.com/trailofbits/incorrect_getsigners/testutil/sample"
)

func TestMsgDelegate_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDelegate
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDelegate{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDelegate{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
